curso_django_2_practico_desarrollo_web_python_3/webpersonal/requirements.txt
Pillow==5.0.0
django==2.2.3
django-ckeditor==5.4.0